+++ packages/blade/src/components/ActionList/componentIds.ts
@@ -1,9 +1,5 @@
 export const componentIds = {
   ActionList: 'ActionList',
-  ActionListHeader: 'ActionListHeader',
-  ActionListHeaderIcon: 'ActionListHeaderIcon',
-  ActionListFooter: 'ActionListFooter',
-  ActionListFooterIcon: 'ActionListFooterIcon',
   ActionListItem: 'ActionListItem',
   ActionListItemAsset: 'ActionListItemAsset',
   ActionListItemIcon: 'ActionListItemIcon',