+++ packages/blade/src/components/BottomSheet/__tests__/BottomSheet.native.test.tsx
@@ -17,7 +17,7 @@ describe('<BottomSheet />', () => {
             title="Address Details"
             subtitle="Saving addresses will improve your checkout experience"
             trailing={<Badge variant="positive">Action Needed</Badge>}
-            titleSuffix={<Counter intent="positive" value={2} />}
+            titleSuffix={<Counter variant="positive" value={2} />}
           />
           <BottomSheetFooter>
             <Button isFullWidth variant="secondary">
@@ -60,7 +60,7 @@ describe('<BottomSheet />', () => {
       );
     };
     expect(() => renderWithTheme(<Example />)).toThrow(
-      '[Blade Header]: Only one of `Button, Badge, Link, Text` component is accepted as trailing',
+      '[Blade Header]: Only one of `Button, Badge, Link, Text, Amount` component is accepted as trailing',
     );
     mockConsoleError.mockRestore();
   });