+++ packages/blade/src/components/Dropdown/index.ts
@@ -2,3 +2,5 @@ export * from './Dropdown';
 export * from './DropdownOverlay';
 export * from './DropdownButton';
 export * from './DropdownLink';
+export * from './DropdownHeaderFooter';
+export * from './types';