+++ packages/blade/src/components/Tooltip/__tests__/Tooltip.web.test.tsx
@@ -32,6 +32,7 @@ describe('<Tooltip />', () => {
     // snapshot while on opened
     fireEvent.focus(getByRole('button', { name: buttonText }));
     expect(screen.queryByRole('tooltip')).toBeInTheDocument();
+    expect(screen.queryByRole('tooltip')).toHaveStyle({ 'z-index': 1100 });
     expect(container).toMatchSnapshot();
   });